+++ app/src/container/SearchContainer/reducer.js
@@ -2,15 +2,23 @@ const initialState = {
   searchString: '',
   data: [],
   filteredData: [],
-  showParkspots: true,
+  showLocations: false,
   isLoading: false,
   filters: [],
   favourites: [],
+  lastSearches: [],
   selectedLocation: null
 };
 
 export default function (state: any = initialState, action: Function) {
   if (action.type === 'UPDATE_SEARCH_STRING') {
+    if (action.searchString === "") {
+      return {
+        ...state,
+        showLocations: false,
+        searchString: action.searchString,
+      };
+    }
     return {
       ...state,
       searchString: action.searchString,
@@ -37,10 +45,17 @@ export default function (state: any = initialState, action: Function) {
       isLoading: false,
     };
   } else if (action.type === 'FETCH_LOCATIONS_SUCCESS') {
+    action.locations.forEach(location => {
+      state.favourites.find((fav) => {
+        if (fav.id === location.id) {
+          location.favourite = true;
+        }
+      });
+    })
     return {
       ...state,
       data: action.locations,
-      showParkspots: false,
+      showLocations: true,
       isLoading: false,
     };
   } else if (action.type === 'FILTER_DATA') {
@@ -81,34 +96,65 @@ export default function (state: any = initialState, action: Function) {
     };
   }
 
-  //checks if newFav is element of favourites - if not yet, includes it
+  //checks if fav is element of favourites - if not yet, includes it
   else if (action.type === 'ADD_FAVOURITE') {
-    if (state.favourites.includes(action.newFav)) {
+    var found = state.favourites.find((fav) => {
+      return fav.id === action.fav.id;
+    });
+    state.lastSearches.find((fav) => {
+      if ( fav.id === action.fav.id ) {
+        fav.favourite = true;
+        return true;
+      }
+    });
+    if (found) {
       return {
         ...state,
       };
-    }
-    else {
+    } else {
+      action.fav.favourite = true;
       return {
         ...state,
-        favourites: [...state.favourites, action.newFav]
+        favourites: [...state.favourites, action.fav]
       };
     }
   }
 
-  //checks if remFav is element of favourites - if so, removes it
+  //checks if fav is element of favourites - if so, removes it
   else if (action.type === 'REM_FAVOURITE') {
-    if (state.favourites.includes(action.remFav)) {
+    const newFavourites = state.favourites;
+    state.favourites.find((fav) => {
+      if (fav.id === action.fav.id) {
+        action.fav.favourite = false;
+        fav.favourite = false;
+        newFavourites = state.favourites.filter(item => item.id !== action.fav.id);
+      }
+      return fav.id === action.fav.id
+    });
+    state.lastSearches.find((fav) => {
+      if ( fav.id === action.fav.id ) {
+        fav.favourite = false;
+        return true;
+      }
+    });
+    return {
+      ...state,
+      favourites: [...newFavourites]
+    };
+  }
 
-      const newFavourites = state.favourites.filter(item => item.id !== action.remFav.id);
+  else if (action.type === 'ADD_LAST_SEARCHED') {
+    var found = state.lastSearches.find((place) => {
+      return place.id === action.place.id;
+    });
+    if (found) {
       return {
         ...state,
-        favourites: [...newFavourites]
       };
-    }
-    else {
+    } else {
       return {
         ...state,
+        lastSearches: [...state.lastSearches, action.place]
       };
     }
   }